// src/types/candidate.ts
export type CandidateProfile = {
  source?: string;
  jobTitle?: string;
  jobLocation?: string;
  fullApplicationUrl?: string;
  availability?: string;
  interviewAvailability?: string;
  currentSalary?: string;   
  expectedSalary?: string; 
  saferedirectUrl?: string;
  resumeUrl?: string;
  cvLink?: string;
  applied_role?: string;
  resumeText?: string;
  age?: string;
  candidate_name?: string;
  candidate_email?: string;
};